pyproject.toml
[tool.poetry]
name = "intelli-crawler"
version = "0.1.0"
description = "Intelli-Crawler: 智能适应型命令行爬虫工具"
authors = ["IntelliCrawler Team <dev@example.com>"]
readme = "README.md"
packages = [{ include = "intelli_crawler" }]

[tool.poetry.scripts]
intelli-crawler = "intelli_crawler.app:cli"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
httpx = "^0.27.0"
pydantic = "^2.7.0"
PyYAML = "^6.0.1"
structlog = "^24.1.0"
tqdm = "^4.66.0"
rich = "^13.7.0"
APScheduler = "^3.10.4"
selectolax = "^0.3.15"
pymongo = "^4.7.2"
cryptography = "^42.0.5"
python-json-logger = "^2.0.7"
playwright = { version = "^1.44.0", optional = true }

[tool.poetry.extras]
playwright = ["playwright"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
respx = "^0.20.2"
responses = "^0.25.0"
black = "^24.4.0"
ruff = "^0.4.6"
mypy = "^1.10.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "Q", "UP", "B"]
ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
implicit_reexport = false
